pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "fastembed>=0.7.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.5.2",
    "langsmith>=0.4.5",
    "openai>=1.95.1",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.13.2",
    "pymupdf>=1.26.3",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.14.3",
    "secure-smtplib>=0.1.1",
    "uvicorn>=0.35.0",
]
